added 'games' and 'user'  subdirectories in db directory for organization. added files to /db/games for Genres model, Publishers model, and Platforms model
