+++ models/games/Games.js
@@ -1,0 +1,27 @@
+const { Model, DataTypes } = require("sequelize");
+const Sequelize = require("../config/connection");
+
+class Games extends Model {}
+
+Games.init({
+	id: {
+		type: DataTypes.INTEGER,
+		primaryKey: true,
+	},
+	title: {
+		type: DataTypes.STRING,
+		allowNull: false,
+	},
+	publisher: {
+		type: DataTypes.STRING,
+		allowNull: false,
+	},
+	release_date: {
+		type: DataTypes.DATE,
+		allowNull: false,
+	},
+	cover_image: {
+		type: DataTypes.STRING,
+		allowNull: false,
+	},
+});
